Got some logic started for handling the defer stack
I'm not in a position to really map out the best operations flow to
start parsing the file. But I knew about some of the enablers I was
going to need. Commiting this to preserve some of what I was thinking
but I think I need to write out how I want to go about this