Calling this 0.4
It's working on all 4 of my test files. (I don't know why I didn't tag v0.3)

The notable change here was a discovery in the way return statements could be
handled. The return was written as part of a loop within the deferstack. This
obviously doesn't make sense. Fixing that seemed to yeild positive results in
all my tests so far.

There are SEVERAL more complex tests that could be written to further validate
this function. Next steps MAY be to clean up the parsing as it's rather
confusing and I do seek to make this project public once I feel it could be a
resume builder