Removing misleading files.
One is a backup that was conveninet to reference instad of playing with git
realtime... should NEVER have been pushed.

The other is a test file. This should be noted as a commit to return to.
I have my visual tests. But, amidst refactoring... I didn't follow up on
the tests. But visual tests indicate this repo may be ready for public
consumption... and I don't want these tests that are faling due to not
being maintained alongside structural changes to poorly reflect on the
status of the repo.